e-modules/comp-scale: disable handling of add.src event for now
this works in usual case. though sometimes a src.add comes in where it shouldnt
with a src that seems not to be appropriate. i.e. not cw->native while tfp
is active and window is not shaped. it segv then in mirror add. need to figure 
out whats going on there


git-svn-id: svn+ssh://svn.enlightenment.org/var/svn/e/trunk/E-MODULES-EXTRA/comp-scale@57524 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33